requirements.txt
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.1
django-bootstrap-dynamic-formsets==0.5.0
django-heroku==0.3.1
django-storages==1.7.1
dropbox==9.4.0
factory_boy==2.12.0
factory_djoy==2.1.2
geographiclib==1.49
geopy==1.20.0
gunicorn==19.9.0
idna==2.8
mock==3.0.5
Pillow==6.2.0
psycopg2==2.8.3
pytest==5.3.2
pytest-django==3.7.0
pytest-factoryboy==2.0.3
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.3
